requirements.txt
# LangGraph
langgraph>=0.2.0
langchain-anthropic>=0.1.0
langgraph-checkpoint-sqlite>=3.0.0

# MCP
mcp>=1.0.0

# Webhook server
flask>=3.0.0

# Utilities
python-dotenv>=1.0.0

# Gmail OAuth (Phase 2)
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
